// src/app/modules/common-regobs-api/models/avalanche-eval-problem2-edit-model.ts
/**
 * NOTE: This file is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */



export interface AvalancheEvalProblem2EditModel { 
    /**
     * Skredtype. I appen er dette 1. felt under skredproblem. The AvalancheExtKD unique identifier
     */
    AvalancheExtTID?: number;
    AvalCauseAttributeCrystalTID: number;
    AvalCauseAttributeLightTID: number;
    AvalCauseAttributeSoftTID: number;
    AvalCauseAttributeThinTID: number;
    /**
     * Hvor dypt ligger det overnevnte svake laget? The AvalCauseDepthKD unique identifier
     */
    AvalCauseDepthTID?: number;
    /**
     * Hvilket svakt lag løsner skredet på? The AvalCauseKD unique identifier
     */
    AvalCauseTID?: number;
    /**
     * Sannsynlighet for skred. The AvalProbabilityKD unique identifier
     */
    AvalProbabilityTID?: number;
    /**
     * 
     */
    AvalPropagationTID?: number;
    /**
     * The AvalTriggerSimpleKD unique identifier
     */
    AvalTriggerSimpleTID?: number;
    /**
     * Kommentar til skredproblemet
     */
    Comment?: string;
    /**
     * Sannsynlig tilleggsbelastning for å utløse skred. The DestructiveSizeKD unique identifier
     */
    DestructiveSizeTID?: number;
    /**
     * Øverste høyde på “utsatt høyde” symbolet.
     */
    ExposedHeight1?: number;
    /**
     * Nederste høyde på “utsatt høyde” symbolet.
     */
    ExposedHeight2?: number;
    /**
     * Hvilket symbol brukes? Er utsatt tereng over ExposedHeight2 eller under den? The ExposedHeightComboKD unique identifier
     */
    ExposedHeightComboTID?: number;
    /**
     * Velg utsatte retninger
     */
    ValidExposition?: string;
}